CPM-734: Move Labels into components
